+++ test/test-custom-context.js
@@ -182,7 +182,7 @@ describe('custom context', function() {
   });
 
 
-  describe('should allow retrival of context value', function() {
+  describe('should allow retrieval of context value', function() {
 
     function computedValue(expression, context = {}, dialect = 'feel') {
 
@@ -436,19 +436,200 @@ describe('custom context', function() {
     });
 
 
-    it.skip('function', function() {
+    it('function invocation (single argument)', function() {
+
+      // when
+      const shape = computedValue(`
+        abs(-22)
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {}
+        }
+      });
+    });
+
+
+    it('function invocation (multiple arguments)', function() {
+
+      // when
+      const shape = computedValue(`
+        substring("foobar", 3)
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {}
+        }
+      });
+    });
+
+
+    it('function invocation (no arguments)', function() {
+
+      // when
+      const shape = computedValue(`
+        now()
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {}
+        }
+      });
+    });
+
+
+    it('function invocation (nested)', function() {
+
+      // when
+      const shape = computedValue(`
+        abs(round(-22.5))
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {}
+        }
+      });
+    });
+
+
+    it('function invocation (context-defined, preserves return shape)', function() {
+
+      // when
+      const shape = computedValue(`
+        {
+          a: function() { a+: { b-: 1 } },
+          b: a()
+        }.b
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {
+            'a +': {
+              value: {
+                atomicValue: undefined,
+                entries: {
+                  'b -': {
+                    value: {
+                      atomicValue: 1,
+                      entries: {}
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      });
+    });
+
+
+    it('function invocation (context-defined, nested unknown)', function() {
+
+      // when
+      const shape = computedValue(`
+        {
+          a: function() { a+: { b-: abs(x) } },
+          b: a()
+        }.b
+      `);
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {
+            'a +': {
+              value: {
+                atomicValue: undefined,
+                entries: {
+                  'b -': {
+                    value: {
+                      atomicValue: undefined,
+                      entries: {}
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      });
+    });
+
+
+    it('function invocation (context-defined, nested unknown with provided context)', function() {
+
+      // when
+      const shape = computedValue(`
+        {
+          a: function() { a+: { b-: abs(x) } },
+          b: a()
+        }.b
+      `, { x: -5 });
+
+      // then
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {
+            'a +': {
+              value: {
+                atomicValue: undefined,
+                entries: {
+                  'b -': {
+                    value: {
+                      atomicValue: undefined,
+                      entries: {}
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      });
+    });
+
+
+    it('function definition', function() {
 
       // when
       const shape = computedValue(`
         {
           add: function(a, b) { result: a + b },
           n: add(1, 2)
-        }
+        }.n
       `);
 
       // then
-      // TODO(nikku): support this
-      expect(shape).to.eql();
+      expect(shape).to.eql({
+        value: {
+          atomicValue: undefined,
+          entries: {
+            result: {
+              value: {
+                atomicValue: 1,
+                entries: {}
+              }
+            }
+          }
+        }
+      });
     });
 
   });